adding env variables

+++ _zod/index.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 
 export const reqInput = z.object({
-  source_language: z.number(),
+  source_language: z.number().int(),
   content: z.string(),
-  target_language: z.number(),
+  target_language: z.number().int(),
 });
